Adding support for decommissioning
The pluggable hardware manager should be able be used for decommissioning,
such as erasing disks, updating firmwares, changing BIOS settings, and
anything else specific hardware needs. This works with the agent driver
to store state in Ironic and request out of band reboots (which may be
required to make BIOS/firmware updates work). Also allows the agent to do
some basic verification of the node's properties, to ensure they haven't
changed, such as a stick of memory dying.

Change-Id: Ia42a8e75d2b8887b2351de97c2137fe21a7f4041